Bugs fixes, temperature is not the cpu temp

+++ src-tauri/src/commands.rs
@@ -6,7 +6,7 @@ use crate::*;
 
 #[derive(Serialize, Deserialize)]
 struct Config {
-    temperature_limit: i16,
+    temperature_limit: f32,
 }
 
 
@@ -15,7 +15,7 @@ struct Config {
 
 
 #[tauri::command]
-pub fn set_limit_command(new_limit: i16, app: AppHandle, state: State<'_, AppState>) {
+pub fn set_limit_command(new_limit: f32, app: AppHandle, state: State<'_, AppState>) {
     {
         let mut limit = state.temperature_limit.lock().unwrap();
         *limit = new_limit;
@@ -25,11 +25,11 @@ pub fn set_limit_command(new_limit: i16, app: AppHandle, state: State<'_, AppSta
 }
 
 #[tauri::command]
-pub fn get_limit_command(state: State<'_, AppState>) -> i16 {
+pub fn get_limit_command(state: State<'_, AppState>) -> f32 {
     *state.temperature_limit.lock().unwrap()
 }
 
 #[tauri::command]
-pub fn get_temperature_command(state: State<'_, AppState>) -> i16 {
+pub fn get_temperature_command(state: State<'_, AppState>) -> f32 {
     *state.temperature.lock().unwrap()
 }
